add scripts to validate and compare results, and add expected results on small datasets

+++ benchmarks/tpcds/queries/78.sql
@@ -35,7 +35,7 @@ with ws as
           where sr_ticket_number is null
           group by d_year, ss_item_sk, ss_customer_sk)
 select ss_customer_sk,
-       round(cast(ss_qty as numeric(18,0)) / (coalesce(ws_qty, 0) + coalesce(cs_qty, 0)), 2)::integer ratio,
+       round(cast(ss_qty as numeric(18,0)) / (coalesce(ws_qty, 0) + coalesce(cs_qty, 0)), 2) ratio,
        ss_qty                                                                  store_qty,
        ss_wc                                                                   store_wholesale_cost,
        ss_sp                                                                   store_sales_price,
